\docType{data}
\name{bird}
\alias{bird}
\title{bird dataset}
\format{A data.frame with 6 observations, no NAs
\describe{
\item{y}{number positive}
\item{n}{number positive + number negative}
\item{tx}{treatment 'vac' or 'con'}
}}
\description{
  This data set is the outcome of a vaccination-challenge
  study in chickens. Chickens were randomized to one of six
  cages, and the cages were then randomized to vaccine or
  placebo groups. Some time after vaccination, all subjects
  were challenged with the virulent pathogen. Those
  affected by the disease were recorded as positive.
}
\keyword{datasets}

