\docType{data}
\name{birdm}
\alias{birdm}
\title{birdm dataset}
\format{a data.frame with 6 observations, no NAs
\describe{
\item{y}{number positive}
\item{n}{number positive + number negative}
\item{tx}{treatment 'vac' or 'con'}
}}
\description{
  This is an artificial modification of the
  \code{\link{bird}} data set for an example described in
  the PF Vignette, Section 4.2.2.
}
\seealso{
  \code{\link{bird}}
}
\keyword{datasets}

