\docType{data}
\name{Set1}
\alias{Set1}
\title{Set1 dataset}
\format{a data.frame with 6 observation of the following 4 variables, no NAs
\describe{
\item{y}{number positive}
\item{n}{total number in group \code{tx} x \code{clus}}
\item{tx}{treatment 'vac' or 'con'}
\item{clus}{cluster ID}
}}
\description{
  Set1 dataset
}
\references{
  We need some references
}
\keyword{datasets}

