\docType{data}
\name{Table6}
\alias{Table6}
\title{Table6 dataset}
\format{a data.frame with 8 observations of the following variables, no NAs
\describe{
\item{y}{number positive}
\item{n}{number positive + number negative}
\item{tx}{treatment 'a' or 'b'}
\item{clus}{cluster ID}
}}
\description{
  These are the data from Table 1 in Gart (1985). They show
  the number of pulmonary tumors among four combinations of
  sex and strain of mice.
}
\references{
  Gart JJ, 1985. Approximate tests and interval estimation
  of the common relative risk in the combination of \eqn{2
  x 2} tables. \emph{Biometrika} 72:673-677. Table 1.
}
\keyword{datasets}

