% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRtosst.r
\name{.rr.score.asymp}
\alias{.rr.score.asymp}
\title{Internal function.}
\usage{
.rr.score.asymp(y, alpha = 0.05, iter.max = 18, converge = 1e-04, mn = FALSE)
}
\arguments{
\item{y}{data}

\item{alpha}{alpha}

\item{iter.max}{maximum number of iterations}

\item{converge}{convergence criterion}

\item{mn}{boolean whether to calculate MN or use default value of 1.0}
}
\description{
Internal function.
}
\examples{
# none
}
