\docType{data}
\name{New}
\alias{New}
\title{New dataset}
\format{a data frame with 52 observations of the following 3 variables, no NAs
\describe{
\item{cage}{Cage ID. 1 - 26}
\item{tx}{Treatment. One of 'con' or 'vac'}
\item{pos}{Response indicator. 0 = negative or 1 = positive}
}}
\description{
  This data set is the outcome of a vaccination-challenge
  study. Each cage housed two dogs, one randomized to
  vaccine, the other received placebo. All subjects were
  later challenged with the virulent pathogen. Those
  affected by the disease were recorded as positive.
}
\keyword{datasets}

