% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{print.rr1}
\alias{print.rr1}
\alias{print.rror}
\alias{print.rrsi}
\alias{print.rrmp}
\alias{print.rrsc}
\alias{print.rrstr}
\title{Print values for PF data obhects.}
\usage{
\method{print}{rr1}(x, ...)

\method{print}{rror}(x, ...)

\method{print}{rrsi}(x, ...)

\method{print}{rrmp}(x, ...)

\method{print}{rrsc}(x, ...)

\method{print}{rrstr}(x, ...)
}
\arguments{
\item{x}{object of class rr1, rror, rrsi, rrmp, rrstr, rrsc}

\item{...}{other arguments not used by this method}
}
\description{
Print values for PF data obhects.
}
