\name{rat}
\alias{rat}
\title{rat dataset}
\format{a data.frame with 32 observations of the following variables, no NAs
\describe{
\item{y}{number positive}
\item{n}{number positive + number negative}
\item{group}{treatment group: 'control' or 'treated'}
}}
\description{
  Weil's rat data. Two groups of 16 pregnant rats were fed
  a treated diet or a control diet. For each of the 32
  litters, the numbers of neonates weaned are recorded as
  positive.
}
\references{
  Weil's rat data, as shown in Table 1 of Rao and Scott:
  \cr \cr Rao JNK, Scott AJ, 1992. A simple method for the
  analysis of clustered binary data. \emph{Biometrics}
  48:577-585.
}
\keyword{datasets}

