% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{rr1-class}
\alias{rr1-class}
\alias{rr1}
\title{Data class rr1}
\description{
Data class rr1
}

\section{Fields}{

\itemize{
\item \code{estimate}:   vector with point and interval estimate
\item \code{estimator}:   either \code{"PF"} or \code{"IDR"}
\item \code{Y}: data.frame with restructured input
\item \code{rnd}: how many digits to round display
\item \code{alpha}: complement of c.i.
}
}

\seealso{
\link{IDRsc}, \link{RRotsst}, \link{RRtosst}
}
\author{
\link{PF-package}
}
