\name{rr1}
\alias{rr1}
\title{Data class rr1}
\description{
  Data class rr1
}
\section{Fields}{
  \itemize{ \item[\code{estimate}]{ vector with point and
  interval estimate} \item[\code{estimator}]{ either
  \code{"PF"} or \code{"IDR"}} \item[\code{y}]{ data
  vector} \item[\code{rnd}]{ how many digits to round
  display} \item[\code{alpha}]{ complement of c.i.} }
}
\author{
  Marie Vendettuoli
  \email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
  \code{\link{IDRsc}, \link{RRotsst}, \link{RRtosst}}
}

