% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRlsi.r
\name{RRlsi}
\alias{RRlsi}
\title{RR likelihood support interval.}
\usage{
RRlsi(
  y = NULL,
  formula = NULL,
  data = NULL,
  vac_grp = "vac",
  con_grp = "con",
  alpha = 0.05,
  k = 8,
  use.alpha = FALSE,
  pf = TRUE,
  iter.max = 50,
  converge = 1e-06,
  rnd = 3,
  start = NULL,
  track = FALSE,
  full.track = FALSE,
  compare = deprecated()
)
}
\arguments{
\item{y}{Data vector \code{c(y1, n1, y2, n2)} where \code{y} are the positives, \code{n} are
the total, and group 1 is compared to group 2 (control or reference group).}

\item{formula}{Formula of the form \code{cbind(y, n) ~ x}, where y is the number
positive, n is the group size, x is a factor with two levels of treatment.}

\item{data}{data.frame containing variables of formula.}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{alpha}{Complement of the confidence level (see details).}

\item{k}{Likelihood ratio criterion.}

\item{use.alpha}{Base choice of k on its relationship to alpha?}

\item{pf}{Estimate \emph{RR} or its complement \emph{PF}?}

\item{iter.max}{Maximum number of iterations}

\item{converge}{Convergence criterion}

\item{rnd}{Number of digits for rounding. Affects display onlyRR, not
estimates.}

\item{start}{Optional starting value.}

\item{track}{Verbose tracking of the iterations?}

\item{full.track}{Verbose tracking of the iterations?}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}  Text vector stating the factor
levels: \code{compare[1]} is the vaccinate group to which \code{compare[2]} (control
or reference) is compared.}
}
\value{
An object of class \link{rrsi} with the following fields: \code{estimate}:
matrix of point and interval estimates - see details \code{estimator}: either
\code{"PF"} or \code{"RR"} \code{y}: data.frame with "y1", "n1", "y2", "n2" values. \code{rnd}:
how many digits to round the display \code{k}: likelihood ratio criterion
\code{alpha}: complement of confidence level
}
\description{
likelihood support interval for the risk ratio or prevented
fraction by the likelihood profile.
}
\details{
Estimates a likelihood support interval for \emph{RR} or \emph{PF}
by the profile likelihood method using the DUD algorithm.

Likelihood support intervals are usually formed based on the desired
likelihood ratio, often 1 / 8 or 1 / 32. Under some conditions the log
likelihood ratio may follow the chi square distribution. If so,
then \eqn{\alpha = 1 - F(2log(k), 1)}, where \eqn{F} is a chi-square CDF. if
\code{use.alpha = TRUE},  \code{RRlsi()} will make the conversion from \eqn{\alpha} to
\eqn{k}

The data may also be a matrix. In that case \code{Y} would be entered as

\code{matrix(c(y1, n1-y1, y2, n2-y2), 2, 2, byrow = TRUE)}.
}
\examples{
# All examples represent the same observation, with data entry by vector,
# matrix, and formula+data notation.

y_vector <- c(4, 24, 12, 28)
RRlsi(y_vector)

y_matrix <- matrix(c(4, 20, 12, 16), 2, 2, byrow = TRUE)
y_matrix

RRlsi(y_matrix)

require(dplyr)
data1 <- data.frame(group = rep(c("treated", "control"), each = 2),
  y = c(1, 3, 7, 5),
  n = c(12, 12, 14, 14),
  cage = rep(paste("cage", 1:2), 2))

data2 <- data1 |>
  group_by(group) |>
  summarize(sum_y = sum(y),
    sum_n = sum(n))
RRlsi(data = data2, formula =  cbind(sum_y, sum_n) ~ group,
      vac_grp = "treated", con_grp = "control")

}
\references{
Royall R. \emph{Statistical Evidence: A Likelihood Paradigm}. Chapman
& Hall, Boca Raton, 1997.  Section 7.6

Ralston ML, Jennrich RI, 1978. DUD, A Derivative-Free Algorithm for
Nonlinear Least Squares. \emph{Technometrics} 20:7-14.
}
\author{
\link{PF-package}
}
