\name{rrmp}
\alias{rrmp}
\title{Data class rrmp}
\description{
  Data class rrmp
}
\section{Fields}{
  \itemize{ \item[\code{estimate}]{ vector with point and
  interval estimate} \item[\code{estimator}]{ either
  \code{"PF"} or \code{"IDR"}} \item[\code{y}]{ data
  vector} \item[\code{rnd}]{ how many digits to round
  display} \item[\code{alpha}]{ complement of c.i.}
  \item[\code{xtable}]{ 2 x 2 data matrix}
  \item[\code{compare}]{ text vector, same as input}
  \item[\code{frecvec}]{ data arrayed a vector of length 4}
  \item[\code{multvec}]{ data.frame showing the multinomial
  representation of the data} }
}
\author{
  Marie Vendettuoli
  \email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
  \code{\link{RRmpWald}}
}

