\name{rror}
\alias{rror}
\title{Data class rror}
\description{
  Data class rror
}
\section{Fields}{
  \itemize{ \item[\code{estimate}]{ vector with point and
  interval estimate} \item[\code{estimator}]{ either
  \code{"PF"} or \code{"IDR"}} \item[\code{y}]{ data
  vector} \item[\code{rnd}]{ how many digits to round
  display} \item[\code{alpha}]{ complement of c.i.}
  \item[\code{norm}]{ logical indicating Gaussian or t
  interval} \item[\code{degf}]{ degrees of freedom}
  \item[\code{mu}]{ matrix with rows giving probability
  estimates for each of the groups} }
}
\author{
  Marie Vendettuoli
  \email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
  \code{\link{RRor}}
}

