% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRotsst.r
\name{RRotsst}
\alias{RRotsst}
\title{RR exact CI, OTSST method.}
\usage{
RRotsst(
  y = NULL,
  data = NULL,
  formula = NULL,
  vac_grp = "vac",
  con_grp = "con",
  alpha = 0.05,
  pf = TRUE,
  stepstart = 0.1,
  iter.max = 36,
  converge = 1e-06,
  rnd = 3,
  trace.it = FALSE,
  nuisance.points = 120,
  gamma = 1e-06,
  compare = deprecated()
)
}
\arguments{
\item{y}{Data vector c(y1, n1, y2, n2) where y are the positives, n are the
total, and group 1 is compared to group 2 (control or reference).}

\item{data}{data.frame containing variables of the formula.}

\item{formula}{Formula of the form cbind(y, n) ~ x, where y is the number
positive, n is the group size, x is a factor with two levels of treatment.}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{alpha}{Complement of the confidence level.}

\item{pf}{Estimate \emph{RR} or its complement \emph{PF}?}

\item{stepstart}{starting interval for step search}

\item{iter.max}{Maximum number of iterations}

\item{converge}{Convergence criterion}

\item{rnd}{Number of digits for rounding. Affects display only, not
estimates.}

\item{trace.it}{Verbose tracking of the iterations?}

\item{nuisance.points}{number of points over which to evaluate nuisance
parameter}

\item{gamma}{parameter for Berger-Boos correction (restricts range of
nuisance parameter evaluation)}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}  Text vector stating the factor levels: \code{compare[1]} is the
vaccinate group to which \code{compare[2]} (control or reference) is compared.}
}
\value{
An object of class \link{rr1} with the following fields:
\itemize{
\item \code{estimate}: vector with point and interval estimate
\item \code{estimator}: either \code{"PF"} or \code{"RR"}
\item \code{y}: \code{data.frame} with "y1", "n1", "y2", "n2" values.
\item \code{rnd}: how many digits to round the display
\item \code{alpha}: complement of confidence level
}
}
\description{
Estimates confidence interval for the risk ratio or prevented
fraction, exact method based on the score statistic (inverts one two-sided
test).
}
\details{
Estimates confidence intervals based on the score statistic that are
'exact' in the sense of accounting for discreteness. The score statistic is
used to select tail area tables, and the binomial probability is estimated
over the tail area by taking the maximum over the nuisance parameter.
Algorithm is a simple step search.

The data may also be a matrix. In that case \code{Y} would be entered as

\code{matrix(c(y1, n1 - y1, y2, n2 - y2), 2, 2, byrow = TRUE)}.
}
\examples{
# All examples represent the same observation, with data entry by multiple
# options.

y_vector <- c(4, 24, 12, 28)
RRotsst(y_vector, rnd = 3)

y_matrix <- matrix(c(4, 20, 12, 16), 2, 2, byrow = TRUE)
RRotsst(y_matrix, rnd = 3)

require(dplyr)
data1 <- data.frame(group = rep(c("treated", "control"), each = 2),
  y = c(1, 3, 7, 5),
  n = c(12, 12, 14, 14),
  cage = rep(paste("cage", 1:2), 2))

data2 <- data1 |>
  group_by(group) |>
  summarize(sum_y = sum(y),
    sum_n = sum(n))
RRotsst(data = data2, formula =  cbind(sum_y, sum_n) ~ group,
   vac_grp = "treated", con_grp = "control")
}
\references{
Koopman PAR, 1984. Confidence intervals for the ratio of two
binomial proportions. \emph{Biometrics} 40:513-517.

Agresti A, Min Y, 2001.  On small-sample confidence intervals for
parameters in discrete distribution. \emph{Biometrics} 57: 963-971.

Berger RL, Boos DD, 1994. P values maximized over a confidence set for the
nuisance parameter. \emph{Journal of the American Statistical Association}
89:214-220.
}
\seealso{
\link{RRtosst}, \link{rr1}.
}
\author{
\link{PF-package}
}
