% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRsc.r
\name{RRsc}
\alias{RRsc}
\title{RR score based asymptotic CI.}
\usage{
RRsc(
  y = NULL,
  data = NULL,
  formula = NULL,
  vac_grp = "vac",
  con_grp = "con",
  alpha = 0.05,
  pf = TRUE,
  trace.it = FALSE,
  iter.max = 18,
  converge = 1e-06,
  rnd = 3,
  compare = deprecated()
)
}
\arguments{
\item{y}{Data vector c(y1, n1, y2, n2) where y are the positives, n are the
total, and group 1 is compared to group 2 (control or reference group).}

\item{data}{data.frame containing variables of formula.}

\item{formula}{Formula of the form \code{cbind(y, n) ~ x}, where y is the number
positive, n is the group size, x is a factor with two levels of treatment.}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{alpha}{Complement of the confidence level.}

\item{pf}{Estimate \emph{RR} or its complement \emph{PF}?}

\item{trace.it}{Verbose tracking of the iterations?}

\item{iter.max}{Maximum number of iterations}

\item{converge}{Convergence criterion}

\item{rnd}{Number of digits for rounding. Affects display only, not
estimates.}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}  Text vector stating the factor
levels: \code{compare[1]} is the vaccinate group to which \code{compare[2]} (control
or reference) is compared.}
}
\value{
A \link{rrsc} object with the following fields.
\itemize{
\item \code{estimate}: matrix of point and interval estimates - see details
\item \code{estimator}: either \code{"PF"} or \code{"RR"}
\item \code{y}: data.frame with "y1", "n1", "y2", "n2" values.
\item \code{rnd}: how many digits to round the display
\item \code{alpha}: complement of confidence level
}
}
\description{
Estimates confidence intervals for the risk ratio or prevented
fraction based on the score statistic.
}
\details{
Estimates are returned for three estimators based on the score
statistic. The score method was introduced by Koopman (1984). Gart and
Nam's modification (1988) includes a skewness correction. The method of
Miettinen and Nurminen (1985) is a version made slightly more conservative
than Koopman's by including a factor of \eqn{(N - 1) / N}. The starting
estimate for the DUD algorithm is obtained by the modified Katz method (log
method with 0.5 added to each cell). Both forms of the Katz estimate may be
retrieved from the returned object using \code{RRsc()$estimate}.

The data may also be a matrix. In that case \code{Y} would be entered as

\code{matrix(c(y1, n1-y1, y2, n2-y2), 2, 2, byrow = TRUE)}.
}
\examples{
# All examples represent the same observation, with data entry by using
# multiple notation options.

y_vector <- c(4, 24, 12, 28)
RRsc(y_vector)

y_matrix <- matrix(c(4, 20, 12, 16), 2, 2, byrow = TRUE)

RRsc(y_matrix)

require(dplyr)
data1 <- data.frame(group = rep(c("treated", "control"), each = 2),
  y = c(1, 3, 7, 5),
  n = c(12, 12, 14, 14),
  cage = rep(paste("cage", 1:2), 2))

data2 <- data1 |>
  group_by(group) |>
  summarize(sum_y = sum(y),
    sum_n = sum(n))
RRsc(data = data2, formula = cbind(sum_y, sum_n) ~ group,
  vac_grp = "treated", con_grp = "control")

}
\references{
Gart JJ, Nam J, 1988. Approximate interval estimation of the
ratio of binomial parameters: a review and corrections for skewness.
\emph{Biometrics} 44:323-338.

Koopman PAR, 1984. Confidence intervals for the ratio of two binomial
proportions. \emph{Biometrics} 40:513-517.

Miettinen O, Nurminen M, 1985. Comparative analysis of two rates.
\emph{Statistics in Medicine} 4:213-226.

Ralston ML, Jennrich RI, 1978. DUD, A Derivative-Free Algorithm for
Nonlinear Least Squares.
\emph{Technometrics} 20:7-14.
}
\seealso{
\link{rrsc}
}
\author{
\link{PF-package}
}
