\name{rrsi}
\alias{rrsi}
\title{Data class rrsi}
\description{
  Data class rrsi
}
\section{Fields}{
  \itemize{ \item[\code{y}]{ numeric data vector }
  \item[\code{k}]{ likelihood ratio criterion}
  \item[\code{rnd}]{ digits to round display}
  \item[\code{alpha}]{ complement of c.i.}
  \item[\code{estimate}]{ vector with point and interval
  estimate} \item[\code{estimator}]{ either \code{"PF"} or
  \code{"IDR"}} }
}
\author{
  Marie Vendettuoli
  \email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
  \code{\link{IDRlsi}, \link{RRlsi}}
}

