% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRStr.r
\name{RRstr}
\alias{RRstr}
\title{Gart-Nam method, CI for common RR over strata or clusters.}
\usage{
RRstr(
  formula = NULL,
  data = NULL,
  vac_grp = "vac",
  con_grp = "con",
  Y,
  alpha = 0.05,
  pf = TRUE,
  trace.it = FALSE,
  iter.max = 24,
  converge = 1e-06,
  rnd = 3,
  multiplier = 0.7,
  divider = 1.1,
  compare = deprecated()
)
}
\arguments{
\item{formula}{Formula of the form \code{cbind(y, n) ~ x + cluster(w)}, where y is
the number positive, n is the group size, x is a factor with two levels of
treatment, and w is a factor indicating the clusters.}

\item{data}{data.frame containing variables of formula}

\item{vac_grp}{The name of the vaccinated group.}

\item{con_grp}{The name of the control group.}

\item{Y}{Matrix of data. Each row is a stratum or cluster. The columns are
y2, n2, y1, n1. If data entered by formula and dataframe, Y is generated
automatically.}

\item{alpha}{Size of the homogeneity test and complement of the confidence
level.}

\item{pf}{Estimate \emph{RR} or its complement \emph{PF}?}

\item{trace.it}{verbose tracking of the iterations?}

\item{iter.max}{Maximum number of iterations}

\item{converge}{Convergence criterion}

\item{rnd}{Number of digits for rounding. Affects display only, not
estimates.}

\item{multiplier}{internal control parameter for algorithm}

\item{divider}{internal control parameter for algorithm}

\item{compare}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}  Text vector stating the factor
levels: \code{compare[1]} is the control or reference group to which
\code{compare[2]} is compared}
}
\value{
A \link{rrstr} object with the following fields:
\itemize{
\item \code{estimate}: matrix of point and interval estimates - starting value, MLE,
and skewness corrected
\item \code{hom}: list of homogeneity statistic, p-value, and degrees of freedom, or
error message if appropriate.
\item \code{estimator}: either \code{"PF"} or \code{"RR"}
\item \code{y}: \code{data.frame} of restructured input
\item \code{compare}: groups compared
\item \code{rnd}: how many digits to round the display
\item \code{alpha}: size of test; complement of confidence level
}
}
\description{
Estimates confidence intervals for the risk ratio or prevented
fraction from clustered or stratified data.
}
\details{
Uses the DUD algorithm to estimate confidence intervals by the
method of Gart.
}
\note{
Vignette \emph{Examples for Stratified Designs} forthcoming with more
examples.

Call to this function may be one of two formats: (1) specify \code{data} and
\code{formula} or (2) as a matrix \code{Y}

\code{RRstr(formula, data, vac_grp = "b", con_grp = "a", pf = TRUE, alpha = 0.05, trace.it = FALSE, iter.max = 24, converge = 1e-6, rnd = 3, multiplier = 0.7, divider = 1.1)}

\code{RRstr(Y, vac_grp = "b", con_grp = "a", pf = TRUE, alpha = 0.05, trace.it = FALSE, iter.max = 24, converge = 1e-6, rnd = 3, multiplier = 0.7, divider = 1.1)}
}
\examples{
## Table 1 from Gart (1985)
##  as data frame
## "b" is control group
RRstr(cbind(y, n) ~ tx + cluster(clus),
      Table6, vac_grp = "a", con_grp = "b", pf = FALSE)

## or as matrix
RRstr(Y = table6, pf = FALSE)

tst <- data.frame(y = c(0, 2, 0, 4, 0, 3, 0, 7),
n = rep(10, 8),
tx = rep(c("a", "b"), 4),
clus = rep(paste("Row", 1:4, sep = ""), each = 2))

}
\references{
Gart JJ, 1985. Approximate tests and interval estimation of the
common relative risk in the combination of \eqn{2 x 2} tables.
\emph{Biometrika} 72:673-677.

Gart JJ, Nam J, 1988. Approximate interval estimation of the ratio of
binomial parameters: a review and corrections for skewness.
\emph{Biometrics} 44:323-338.

Ralston ML, Jennrich RI, 1978. DUD, A Derivative-Free Algorithm for
Nonlinear Least Squares. \emph{Technometrics} 20:7-14.
}
\seealso{
\link{rrstr}
}
\author{
\link{PF-package}
}
