% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsbWt.r
\name{rsbWt}
\alias{rsbWt}
\title{Rao-Scott weighting.}
\usage{
rsbWt(fit = NULL, subset.factor = NULL, fit.only = TRUE)
}
\arguments{
\item{fit}{A \link[stats:glm]{stats::glm} object.}

\item{subset.factor}{Factor for estimating phi by subset.  Will be converted
to a factor if it is not a factor.}

\item{fit.only}{Return only the new fit? If FALSE, also returns the weights
and phi estimates.}
}
\value{
A list with the following elements.
\itemize{
\item \code{fit}: the new model fit, updated by the estimated weights
\item \code{weights}: vector of weights
\item \code{d}: vector of \eqn{{d}_{i}}{d_i} estimates
}
}
\description{
Rao-Scott weighting of clustered binomial observations.
}
\details{
Estimates the cluster design effect \eqn{{d}_{i}}{d_i} as the
variance inflation due to clustering by the method of Rao and Scott.
Observations are then weighted by the inverse of the \eqn{{d}_{i}}{d_i}.
}
\examples{
birdm.fit <- glm(cbind(y, n - y) ~ tx-1, binomial, birdm)
RRor(rsbWt(birdm.fit))
#
# 95\% t intervals on 4 df
#
# PF
#     PF     LL     UL
#  0.479 -1.061  0.868
#
#       mu.hat    LL     UL
# txcon  0.768 0.968 0.2659
# txvac  0.400 0.848 0.0737
#
}
\references{
Rao JNK, Scott AJ, 1992. A simple method for the analysis of
clustered binary data. \emph{Biometrics} 48:577-585.
}
\seealso{
\link{RRor}, \link{rsb}.
}
\author{
\link{PF-package}
}
