% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF-package.r
\docType{data}
\name{bird}
\alias{bird}
\title{bird dataset}
\format{
a data.frame with 6 observations of the following 4 variables, no NAs
\itemize{
\item \code{y}: number positive
\item \code{n}: total number in group \code{tx} x \code{all}
\item \code{tx}: treatment "vac" or "con"
\item \code{all}: all?
}
}
\description{
bird dataset
}
\references{
we need some references
}
\keyword{datasets}
