\docType{data}
\name{bird}
\alias{bird}
\title{bird dataset}
\format{a data.frame with 6 observations of the following 4 variables, no NAs
\describe{
\item{y}{number positive}
\item{n}{total number in group \code{tx} x \code{all}}
\item{tx}{treatment 'vac' or 'con'}
\item{all}{all?}
}}
\description{
  bird dataset
}
\references{
  we need some references
}
\keyword{datasets}

