% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF-package.r
\docType{data}
\name{birdm}
\alias{birdm}
\title{birdm dataset}
\format{
a data.frame with 6 observations of the following 4 variables, no NAs
\describe{

\item{y}{number positive}

\item{n}{total number in group \code{tx} x \code{all}}

\item{tx}{treatment 'vac' or 'con'}

\item{all}{all?}

}
}
\description{
birdm dataset
}
\references{
we need some references
}
\keyword{datasets}
