% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF-package.r
\docType{data}
\name{Set1}
\alias{Set1}
\title{Set1 dataset}
\format{
a data.frame with 6 observation of the following 4 variables, no NAs
\itemize{
\item \code{y}: number positive
\item \code{n}: total number in group \code{tx} x \code{clus}
\item \code{tx}: treatment "vac" or "con"
\item \code{clus}: cluster ID
}
}
\description{
Set1 dataset
}
\references{
We need some references
}
\keyword{datasets}
