\docType{data}
\name{Table6}
\alias{Table6}
\title{Table6 dataset}
\format{a data.frame with 8 observations of the following 4 variables, no NAs
\describe{
\item{y}{number positive}
\item{n}{total number in group \code{tx} x \code{clus}}
\item{tx}{treatment 'a' or 'b'}
\item{clus}{cluster ID}
}}
\description{
  Table6 dataset
}
\references{
  Table 1 from Gart (1985)
}
\keyword{datasets}

