% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF-package.r
\docType{data}
\name{New}
\alias{New}
\title{New dataset}
\format{
a data frame with 52 observations of the following 3 variables, no
NAs
\itemize{
\item \code{cage}: cage ID. 1 - 26
\item \code{tx}: treatment. one of "con" or "vac"
\item \code{pos}: numeric indicator of positive response. 0 = FALSE or 1 = TRUE
}
}
\description{
New dataset
}
\references{
We need some references
}
\keyword{datasets}
