\docType{data}
\name{New}
\alias{New}
\title{New dataset}
\format{a data frame with 52 observations of the following 3 variables, no NAs
\describe{
\item{cage}{cage ID. 1 - 26}
\item{tx}{treatment. one of 'con' or 'vac'}
\item{pos}{numeric indicator of positive response. 0 = FALSE or 1 = TRUE}
}}
\description{
  New dataset
}
\references{
  We need some references
}
\keyword{datasets}

