% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF.r
\docType{package}
\name{PF-package}
\alias{PF}
\alias{PF-package}
\title{PF: Prevented fraction}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions related to PF (prevented fraction). Calculate incidence density ratio, confidence interval, and Rao-Scott weights of PF by multiple methods. See http://goo.gl/eJ6Rxi for definition of PF. No endorsement, claim, or warranty is implied for this package. It is made available for investigational or pedagogical use only. See https://www.aphis.usda.gov/animal_health/vet_biologics/publications/STATWI0007.pdf for futher details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ABS-dev/PF/}
  \item \url{https://abs-dev.github.io/PF/}
  \item Report bugs at \url{https://github.com/ABS-dev/PF/issues/}
}

}
\author{
\strong{Maintainer}: CVB Statistics \email{CVB.Data.Help@usda.gov} [former owner]

Authors:
\itemize{
  \item David Siev (until 2019)
  \item Marie Vendettuoli (until 2019)
  \item Thomas Kent (starting 2020)
}

}
\keyword{internal}
