% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PF-package.r
\docType{package}
\name{PF-package}
\alias{PF-package}
\alias{PF}
\title{Package for PF.}
\description{
Includes functions related to prevented fraction.
}
\details{
\tabular{ll}{
Package: \tab pf-package\cr
Type: \tab Package\cr
Version: \tab 9.5.3\cr
Date: \tab 2018-08-01\cr
License: \tab MIT\cr
LazyLoad: \tab yes\cr
LazyData: \tab yes\cr
}
}
\section{Resources}{


\itemize{
\item GUIDANCE: \url{https://www.aphis.usda.gov/aphis/ourfocus/animalhealth/veterinary-biologics/biologics-regulations-and-guidance/ct_vb_statwi}
\item QUICK START: \url{https://github.com/ABS-dev/PF/blob/master/README.md}
\item BUG REPORTS: \url{https://github.com/ABS-dev/PF/issues}
}
}

\author{
David Siev \email{David.Siev@aphis.usda.gov}
}
