\docType{package}
\name{PF-package}
\alias{PF}
\alias{PF-package}
\title{Package for PF.}
\description{
  Includes functions related to prevented fraction.
}
\details{
  \tabular{ll}{ Package: \tab pf-package\cr Type: \tab
  Package\cr Version: \tab 9.5\cr Date: \tab 2013-08-29\cr
  License: \tab MIT\cr LazyLoad: \tab yes\cr LazyData: \tab
  yes\cr }
}
\examples{
#---------------------------------------------
# Checking PF package
#---------------------------------------------
example(RRsc)
example(RRstr)
example(RRmh)
example(RRor)
example(phiWt)
example(tauWt)
example(rsbWt)
example(rsb)
example(RRlsi)
example(IDRsc)
example(IDRlsi)
#---------------------------------------------
# The next two take a moment to run
#---------------------------------------------
example(RRtosst)
example(RRotsst)
#---------------------------------------------
# End examples
#---------------------------------------------
invisible()
}
\author{
  David Siev \email{David.Siev@aphis.usda.gov}
}

