% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{pf-class}
\alias{pf-class}
\alias{pf}
\title{Data class pf}
\description{
data class pf
}

\section{Fields}{

\itemize{
\item[\code{estimator}]{  either \code{"PF"} or \code{"IDR"}}
\item[\code{rnd}]{  how many digits to round display}
\item[\code{alpha}]{  complement of c.i.}
}
}

\seealso{
\code{\link{rr1}, \link{rrsi}, \link{rrsc}, \link{rrstr}}
}
\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
