\name{pf}
\alias{pf}
\title{Data class pf}
\description{
  Data class pf
}
\section{Fields}{
  \itemize{ \item[\code{estimator}]{ either \code{"PF"} or
  \code{"IDR"}} \item[\code{rnd}]{ how many digits to round
  display} \item[\code{alpha}]{ complement of c.i.} }
}
\author{
  Marie Vendettuoli
  \email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
  \code{\link{rr1}, \link{rrsi}, \link{rrsc}, \link{rrstr}}
}

