% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phiWt.r
\name{phiWt}
\alias{phiWt}
\title{Binomial dispersion parameter.}
\usage{
phiWt(fit, subset.factor = NULL, fit.only = TRUE, show.warns = FALSE)
}
\arguments{
\item{fit}{A \link{glm} object.}

\item{subset.factor}{Factor for estimating phi by subset.  Will be converted
to a factor if it is not a factor.}

\item{fit.only}{Return only the new fit?  If FALSE, also returns the weights
and phi estimates.}

\item{show.warns}{Show warnings}
}
\value{
A list with the following elements. \code{fit}: the new model fit,
updated by the estimated weights \code{weights}: vector of weights \code{phi}: vector
of phi estimates
}
\description{
MME estimate of dispersion parameter phi.
}
\details{
Estimates binomial dispersion parameter \eqn{\phi} by the method of
moments. Refits the model, weighting the observations by \eqn{1/\phi}. Uses
\code{quasibinomial} family in \code{glm()}.
}
\examples{
birdm.fit <- glm(cbind(y, n - y) ~ tx-1, binomial, birdm)
RRor(phiWt(birdm.fit))
#
# 95\% t intervals on 4 df
#
# PF
#     PF     LL     UL
#  0.479 -0.537  0.823
#
#       mu.hat   LL    UL
# txcon  0.768 0.95 0.367
# txvac  0.400 0.78 0.111
#
}
\references{
Wedderburn RWM, 1974. Quasi-likelihood functions, generalized
linear models, and the Gauss-Newton method. \emph{Biometrika} 61:439-447.
}
\seealso{
\link{tauWt}, \link{RRor}.
}
\author{
\link{PF-package}
}
