\name{print.rr1}
\alias{print.rr1}
\alias{print.rrmp}
\alias{print.rror}
\alias{print.rrsc}
\alias{print.rrsi}
\alias{print.rrstr}
\title{Print values for PF data obhects.}
\usage{
  \method{print}{rr1} (x, ...)

  \method{print}{rror} (x, ...)

  \method{print}{rrsi} (x, ...)

  \method{print}{rrmp} (x, ...)

  \method{print}{rrsc} (x, ...)

  \method{print}{rrstr} (x, ...)
}
\arguments{
  \item{x}{object of class rr1, rror, rrsi, rrmp, rrstr,
  rrsc}

  \item{...}{other arguments not used by this method}
}
\description{
  Print values for PF data obhects.
}

