\name{rat}
\alias{rat}
\title{rat dataset}
\format{a data.frame with 32 observations of the following 3 variables, no NAs
\describe{
\item{y}{number positive}
\item{n}{total number}
\item{group}{treatment group: 'control' or 'treated'}
}}
\description{
  rat dataset
}
\references{
  Weil's rat data (Table 1 of Rao and Scott)
}
\keyword{datasets}

