% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{rrmp-class}
\alias{rrmp-class}
\alias{rrmp}
\title{Data class rrmp}
\description{
data class rrmp
}

\section{Fields}{

\itemize{
\item \code{estimate}:  vector with point and interval estimate
\item \code{estimator}:  either \code{"PF"} or \code{"IDR"}
\item \code{Y}:  data vector
\item \code{rnd}:  how many digits to round display
\item \code{alpha}:  complement of c.i.
\item \code{compare}:  text vector, same as input
\item \code{multvec}:  data.frame showing the multinomial representation of the data
}
}

\seealso{
\link{RRmpWald}
}
\author{
\link{PF-package}
}
