% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RRmpWald.r
\name{RRmpWald}
\alias{RRmpWald}
\title{Wald confidence intervals for RR from matched pairs}
\usage{
RRmpWald(
  formula = NULL,
  data = NULL,
  compare = c("vac", "con"),
  affected = 1,
  x,
  alpha = 0.05,
  pf = TRUE,
  tdist = TRUE,
  df = NULL,
  rnd = 3
)
}
\arguments{
\item{formula}{Formula of the form \code{y ~ x + cluster(w)}, where \code{y} is the
indicator for an individual's positive response, \code{x} is a factor with two
levels of treatment, and \code{w} identifies the pairs.}

\item{data}{\code{data.frame} containing variables in formula}

\item{compare}{Text vector stating the factor levels: \code{compare[1]} is the
vaccinate group to which \code{compare[2]} (control or reference) is compared.}

\item{affected}{Indicator for positive response}

\item{x}{Alternative data input. Instead of formula and data frame, data may
be input as frequency vector. See example for how to order this vector.}

\item{alpha}{Complement of the confidence level}

\item{pf}{Estimate \emph{RR} or its complement \emph{PF}?}

\item{tdist}{Use t distribution?}

\item{df}{Degrees of freedom. When NULL, the function will default to `df = N
\itemize{
\item 2`, where N is the total number of pairs.
}}

\item{rnd}{Number of digits for rounding. Affects display only, not
estimates.}
}
\value{
A \link{rrmp} object with the following fields:
\itemize{
\item \code{estimate}: vector of point and interval estimates - see details
\item \code{estimator}: either \code{"PF"} or \code{"RR"}
\item \code{compare}: text vector, same as input
\item \code{alpha}: complement of confidence level
\item \code{rnd}: how many digits to round the display
\item \code{multvec}: data frame showing the multinomial representation of the data
}
}
\description{
Estimates confidence intervals for the risk ratio or prevented
fraction from matched pairs.
}
\details{
Estimates confidence intervals for the risk ratio or prevented
fraction from matched pairs. The response is the tetranomial vector \code{c(11, 12, 21, 22)}, where the first index is the row and the the second index is
the column when displayed as a 2x2 table. Wald type confidence intervals
are found by applying the delta method to the multinomial variance. This
method fails when there are no responders in one of the treatment groups.

Alternative forms of data entry are illustrated by the output, say \code{Y},
where \code{c(Y$xtable) = Y$freqvec = Y$multvec$Freq}.

If \code{RR = 0} (\code{PF = 1}), the function will return degenerate interval.
}
\note{
Experimental functions for estimating profile likelihood intervals are
in the CVBmisc package.

Call to this function may be one of two formats: (1) specify \code{data} and
\code{formula} or (2) as a vector \code{x}

\code{RRmpWald(formula, data, compare = c("vac", "con"), affected = 1, alpha = 0.05, pf = TRUE, tdist = TRUE, df = NULL, rnd = 3)}

\code{RRmpWald(x, compare = c("vac", "con"), affected = 1, alpha = 0, 05, pf = TRUE, tdist = TRUE, df = NULL, rnd = 3)}
}
\examples{
RRmpWald(pos ~ tx + cluster(cage), New, compare = c("vac", "con"))

thistable <- New |>
  tidyr::spread(tx, pos) |>
  dplyr::mutate(vac = factor(vac, levels = 1:0),
    con = factor(con, levels = 1:0)) |>
  with(table(vac, con))
thistable
as.vector(thistable)

RRmpWald(x = as.vector(thistable))
}
\author{
\link{PF-package}
}
