% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{rror-class}
\alias{rror-class}
\alias{rror}
\title{Data class rror}
\description{
data class rror
}

\section{Fields}{


\itemize{

\item[\code{estimate}]{  vector with point and interval estimate}

\item[\code{estimator}]{  either \code{"PF"} or \code{"IDR"}}

\item[\code{y}]{  data vector}

\item[\code{rnd}]{  how many digits to round display}

\item[\code{alpha}]{  complement of c.i.}

\item[\code{norm}]{  logical indicating Gaussian or t interval}

\item[\code{degf}]{  degrees of freedom}

\item[\code{mu}]{  matrix with rows giving probability estimates for each
of the groups}

}
}

\seealso{
\code{\link{RRor}}
}
\author{
\link{PF-package}
}
