% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{rrsc-class}
\alias{rrsc-class}
\alias{rrsc}
\title{Data class rrsc}
\description{
data class rrsc
}

\section{Fields}{

\itemize{
\item[\code{estimate}]{  vector with point and interval estimate}
\item[\code{rnd}]{  how many digits to round display}
\item[\code{alpha}]{  complement of c.i.}
\item[\code{estimator}]{  either \code{"PF"} or \code{"RR"}}
\item[\code{y}]{  data vector}
}
}

\seealso{
\code{\link{RRsc}}
}
\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
