\name{rrsc}
\alias{rrsc}
\title{Data class rrsc}
\description{
  Data class rrsc
}
\section{Fields}{
  \itemize{ \item[\code{estimate}]{ vector with point and
  interval estimate} \item[\code{rnd}]{ how many digits to
  round display} \item[\code{alpha}]{ complement of c.i.}
  \item[\code{estimator}]{ either \code{"PF"} or
  \code{"RR"}} \item[\code{y}]{ data vector} }
}
\author{
  Marie Vendettuoli
  \email{marie.c.vendettuoli@aphis.usda.gov}
}
\seealso{
  \code{\link{RRsc}}
}

