% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{rrsi-class}
\alias{rrsi-class}
\alias{rrsi}
\title{Data class rrsi}
\description{
data class rrsi
}

\section{Fields}{

\itemize{
\item[\code{y}]{ data.frame with restructured input}
\item[\code{k}]{  likelihood ratio criterion}
\item[\code{rnd}]{  digits to round display}
\item[\code{alpha}]{  complement of c.i.}
\item[\code{estimate}]{  vector with point and interval estimate}
\item[\code{estimator}]{  either \code{"PF"} or \code{"IDR"}}
}
}

\seealso{
\code{\link{IDRlsi}, \link{RRlsi}}
}
\author{
\link{PF-package}
}
