% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.r
\docType{class}
\name{rrstr-class}
\alias{rrstr-class}
\alias{rrstr}
\title{Data class rrstr}
\description{
data class rrstr
}

\section{Fields}{

\itemize{
\item[\code{estimate}]{  vector with point and interval estimate}
\item[\code{rnd}]{  how many digits to round display}
\item[\code{alpha}]{  complement of c.i.}
\item[\code{estimator}]{  either \code{"PF"} or \code{"RR"}}
\item[\code{hom}]{  list of homogeneity statistic, p-value, and degrees of freedom}
\item[\code{y}]{  matrix of data}
\item[\code{compare}]{  groups compared}
}
}

\seealso{
\code{\link{RRstr}}
}
\author{
Marie Vendettuoli \email{marie.c.vendettuoli@aphis.usda.gov}
}
