% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsbWt.r
\name{rsb}
\alias{rsb}
\title{Rao-Scott weights.}
\usage{
rsb(y = NULL, n = NULL, formula = NULL, data = NULL, id = NULL)
}
\arguments{
\item{y}{vector of number positive.}

\item{n}{vector of total number.}

\item{formula}{Formula of the form cbind(y, n) ~ id, where y is the number
positive, n is the total number, id is a factor for estimating the weights
by subset.}

\item{data}{data.frame containing variables of formula.}

\item{id}{vector of factor for estimating the weights by subset.}
}
\value{
A list with the following elements. \item{w}{vector of weights}
 \item{d}{vector of \eqn{{d}_{i}}{d_i} estimates}
}
\description{
Rao-Scott weights.
}
\details{
Estimates the cluster design effect \eqn{{d}_{i}}{d_i} as the
 variance inflation due to clustering by the method of Rao and Scott.
 \code{rsb} estimates the \eqn{{d}_{i}}{d_i} for use by \code{rsbWt} or other
 functions.
}
\examples{
# Weil's rat data (Table 1 of Rao and Scott)
rsb(rat$y, rat$n, id = rat$group)$d
#  control  treated
# 1.232495 3.952861
rsb(data = rat, formula = cbind(y, n) ~ group)$d
#  control  treated
# 1.232495 3.952861
}
\references{
Rao JNK, Scott AJ, 1992. A simple method for the analysis of
 clustered binary data. \emph{Biometrics} 48:577-585.
}
\seealso{
\code{\link{rsbWt}}.
}
\author{
\link{PF-package}
}
